[PersistentNodeKeys] Use of system property to get key file format and refactoring
