+++ src/main/resources/application.conf
@@ -2,6 +2,7 @@ etc-client {
 
   client-id = "etc-client"
 
+  //File format of the keys (in plain text) to be set with -Dkey-file-format
   keys-file = "nodeId.keys"
 
   network {
